+++ tests/functional/emit-schema-sdl.ts
@@ -1,5 +1,6 @@
 import "reflect-metadata";
 import fs from "node:fs";
+import asyncFs from "node:fs/promises";
 import path from "node:path";
 import { type GraphQLSchema } from "graphql";
 import shelljs from "shelljs";
@@ -13,7 +14,6 @@ import {
   //  buildSchemaSync,
   defaultPrintSchemaOptions,
 } from "type-graphql";
-import * as filesystem from "@/helpers/filesystem";
 import {
   emitSchemaDefinitionFile,
   emitSchemaDefinitionFileSync,
@@ -89,7 +89,7 @@ describe("Emitting schema definition file", () => {
     });
 
     it("should throw error when unknown error occur while writing file with schema SDL", async () => {
-      jest.spyOn(filesystem, "fsWriteFile").mockRejectedValueOnce({ code: "TEST ERROR" });
+      jest.spyOn(asyncFs, "writeFile").mockRejectedValueOnce({ code: "TEST ERROR" });
       const targetPath = path.join(TEST_DIR, "schemas", "fail1", "schema.graphql");
       let error;
       try {
@@ -102,7 +102,7 @@ describe("Emitting schema definition file", () => {
     });
 
     it("should throw error when unknown error occur while creating directory with schema SDL", async () => {
-      jest.spyOn(filesystem, "fsMkdir").mockRejectedValueOnce({ code: "TEST ERROR" });
+      jest.spyOn(asyncFs, "mkdir").mockRejectedValueOnce({ code: "TEST ERROR" });
       const targetPath = path.join(TEST_DIR, "schemas", "fail2", "schema.graphql");
       let error;
       try {