+++ tests/functional/resolvers.ts
@@ -1,5 +1,6 @@
 /* eslint "@typescript-eslint/no-this-alias": ["error", { "allowedNames": ["self"] }] */
 import "reflect-metadata";
+import { createPubSub } from "@graphql-yoga/subscription";
 import {
   type GraphQLSchema,
   type IntrospectionField,
@@ -28,8 +29,6 @@ import {
   Mutation,
   NoExplicitTypeError,
   ObjectType,
-  PubSub,
-  PubSubEngine,
   Query,
   Resolver,
   type ResolverInterface,
@@ -2179,6 +2178,8 @@ describe("Resolvers", () => {
     let childResolver: any;
     let overrideResolver: any;
 
+    const pubSub = createPubSub();
+
     beforeEach(() => {
       self = null;
     });
@@ -2222,8 +2223,8 @@ describe("Resolvers", () => {
           }
 
           @Mutation(() => Boolean, { name: `${name}Trigger` })
-          async baseTrigger(@PubSub() pubSub: PubSubEngine): Promise<boolean> {
-            await pubSub.publish("baseTopic", null);
+          async baseTrigger(): Promise<boolean> {
+            pubSub.publish("baseTopic", null);
             return true;
           }
 
@@ -2263,8 +2264,8 @@ describe("Resolvers", () => {
         }
 
         @Mutation(() => Boolean)
-        async childTrigger(@PubSub() pubSub: PubSubEngine): Promise<boolean> {
-          await pubSub.publish("childTopic", null);
+        async childTrigger(): Promise<boolean> {
+          pubSub.publish("childTopic", null);
           return true;
         }
       }
@@ -2288,6 +2289,7 @@ describe("Resolvers", () => {
 
       const schemaInfo = await getSchemaInfo({
         resolvers: [childResolver, overrideResolver],
+        pubSub,
       });
       schemaIntrospection = schemaInfo.schemaIntrospection;
       queryType = schemaInfo.queryType;
@@ -2476,6 +2478,71 @@ describe("Resolvers", () => {
       await graphql({ schema, source: query });
 
       expect(self).toBeInstanceOf(childResolver);
+    });
+
+    it("should allow duplicate fieldResolver methods with different schema names for inherited resolvers", async () => {
+      getMetadataStorage().clear();
+      const INHERITED_DYNAMIC_FIELD_NAME_1 = "dynamicallyNamedMethod1";
+      const INHERITED_DYNAMIC_FIELD_NAME_2 = "dynamicallyNamedMethod2";
+
+      const withDynamicallyNamedFieldResolver = (
+        classType: ClassType,
+        BaseResolverClass: ClassType,
+        name: string,
+      ) => {
+        @Resolver(() => classType)
+        class DynamicallyNamedFieldResolver extends BaseResolverClass {
+          @FieldResolver({ name })
+          dynamicallyNamedField(): boolean {
+            return true;
+          }
+        }
+        return DynamicallyNamedFieldResolver;
+      };
+
+      @ObjectType()
+      class SampleObject {
+        @Field()
+        sampleField!: string;
+      }
+
+      @Resolver()
+      class SampleResolver {
+        @Query(() => SampleObject)
+        sampleObject(): SampleObject {
+          return { sampleField: "sampleText" };
+        }
+      }
+
+      const DynamicallyNamedFieldResolver1 = withDynamicallyNamedFieldResolver(
+        SampleObject,
+        SampleResolver,
+        INHERITED_DYNAMIC_FIELD_NAME_1,
+      );
+      const DynamicallyNamedFieldResolver2 = withDynamicallyNamedFieldResolver(
+        SampleObject,
+        DynamicallyNamedFieldResolver1,
+        INHERITED_DYNAMIC_FIELD_NAME_2,
+      );
+
+      const schemaInfo = await getSchemaInfo({
+        resolvers: [DynamicallyNamedFieldResolver2],
+      });
+      schemaIntrospection = schemaInfo.schemaIntrospection;
+      const sampleObjectType = schemaIntrospection.types.find(
+        type => type.name === "SampleObject",
+      ) as IntrospectionObjectType;
+
+      const dynamicField1 = sampleObjectType.fields.find(
+        field => field.name === INHERITED_DYNAMIC_FIELD_NAME_1,
+      )!;
+
+      const dynamicField2 = sampleObjectType.fields.find(
+        field => field.name === INHERITED_DYNAMIC_FIELD_NAME_2,
+      )!;
+
+      expect(dynamicField1).toBeDefined();
+      expect(dynamicField2).toBeDefined();
     });
   });
 });